requirements.txt
torch
torchvision
ultralytics
opencv-python
Pillow
numpy
matplotlib
scikit-learn
albumentations
timm
tensorboard
tqdm
seaborn